Brownfields and national priority sites

A sf data frame of basic information on brownfields and national priority list (superfund) sites in Maryland. This is a subset of data from the Maryland Department of the Environment's (MDE) Land Restoration Program, filtered for sites that are listed as brownfields, NPL sites, or both.

Usage

brownfields_sf

Format

An sf data frame with 150 rows and 11 variables:

id

Integer. An ID, identical to the object ID in the original dataset.

name

Character. Site name listed in the MDE database.

address

Character. Site address(es).

city

Character. Town name.

is_ongoing_assess

Logical, whether assessment of the site is listed as ongoing.

is_ongoing_remed

Logical, whether remediation of the site is listed as ongoing.

is_archived

Logical, whether the site is considered closed.

fy_open

Numeric. Fiscal year cleanup process was opened.

fy_closed

Numeric. Fiscal year cleanup process was closed, if applicable.

site_type

Factor. The site type (brownfield, npl, or both).

geometry

POINT. Location.

Source

Maryland Department of the Environment Land Restoration Program, available at https://mdewin64.mde.state.md.us/LRP/index.html

Examples

 head(brownfields_sf)
#>       id                                             name
#> 1 135066                           FNT Realty Cherry Hill
#> 2 135084                             Bush Valley Landfill
#> 3 135100                               101 Garrett Street
#> 4 135110                      Northhampton, City of Largo
#> 5 135114                          Shellfish Seafood, Ltd.
#> 6 135115 Former PPG - Works No. 7 (Pittsburg Plate Glass)
#>                                    address          city is_ongoing_assess
#> 1          Cherry Hill Road and Metal Road     Baltimore             FALSE
#> 2                                Bush Road      Abingdon             FALSE
#> 3                       101 Garrett Street     Baltimore             FALSE
#> 4 Harry Truman Drive and Mt. Lubentia Road         Largo             FALSE
#> 5                     30660 Hampden Avenue Princess Anne             FALSE
#> 6         11601 Pittsburg Plate Glass Road    Cumberland             FALSE
#>   is_ongoing_remed is_archived fy_open fy_closed  site_type          geometry
#> 1            FALSE       FALSE      NA        NA brownfield -76.6335, 39.2574
#> 2             TRUE       FALSE    1989        NA        npl -76.2619, 39.4671
#> 3            FALSE       FALSE    1994        NA brownfield -76.6143, 39.2376
#> 4            FALSE        TRUE    2003      2003 brownfield -76.8279, 38.8806
#> 5            FALSE       FALSE    2004        NA brownfield -75.6880, 38.2008
#> 6            FALSE        TRUE    1997      2000 brownfield -78.7479, 39.5914