2022-based CPI inflation adjustment factor

A table of inflation adjustment factors for the year 1990 to 2022. This can be used to adjust dollar values for inflation into 2022 dollars, based on annual consumer price index (CPI) values from the Bureau of Labor Statistics. Match your values to cpi by year, then divide your values by the adjustment factor. Because these are annual averages, it won't be appropriate for more granular time periods in the past few years. This adjusts to 2022 dollars to match the spending dataset.

Usage

cpi

Format

A data frame with 33 rows and 2 variables:

year

Numeric. Year of data.

adj_factor22

Numeric. Factor by which to adjust dollar amounts to get 2022-based dollars.

Source

Bureau of Labor Statistics Consumer Price Index for All Urban Consumers (CPI-U) series CUUR0000SA0. https://www.bls.gov/cpi/

Examples

  head(cpi)
#>   year adj_factor22
#> 1 1990        0.446
#> 2 1991        0.465
#> 3 1992        0.479
#> 4 1993        0.494
#> 5 1994        0.506
#> 6 1995        0.521

  # to use cpi, join it to your data and divide values by the inflation factor
  set.seed(1)
  x <- data.frame(year = 2016:2022, value = round(rnorm(7, 100, 10)))
  x |>
    dplyr::left_join(cpi, by = "year") |>
    dplyr::mutate(adj_value = value / adj_factor22)
#>   year value adj_factor22 adj_value
#> 1 2016    94        0.820 114.63415
#> 2 2017   102        0.838 121.71838
#> 3 2018    92        0.858 107.22611
#> 4 2019   116        0.874 132.72311
#> 5 2020   103        0.884 116.51584
#> 6 2021    92        0.926  99.35205
#> 7 2022   105        1.000 105.00000