EPA environmental justice index trend

A dataset containing environmental health risk factors from the EPA's EJSCREEN environment justice index for block groups in Maryland for each year from 2018 to 2023. Values are calculated based on aggregations of risk factors, then given as nationwide percentiles. The column d2_ptile is adjusted for the EPA's two-factor definition of vulneration populations. While different years of this data are formatted differently, all indicators subset here are consistent except for underground storage, which is only available starting in 2021. Note also that block group definitions and their GEOIDs changed with the 2020 decennial census: 2018 to 2021 use 2010 block groups, while 2022 and 2023 use 2020 block groups. If joining with shapefiles from TIGER or elsewhere, you'll need one for 2010 block groups and one for 2020.

Usage

ej_trend

Format

A data frame with 270334 rows and 6 variables:

year

Numeric. Year of data.

bg

Character. The block group FIPS code.

total_pop

Numeric. Total population of the block group that year.

indicator

Factor. The environmental health risk factor, such as proximity to water treatment or air pollution-related cancers.

value_ptile

Integer. The nationwide percentile of indexed values.

d2_ptile

Integer. The percentile of indexed values scaled based on a two-factor demographic index (percent low-income and percent people of color).

Source

Environmental Protection Agency (EPA) EJSCREEN Environment Justice Index. Data portal, definitions, and methodology are available at https://www.epa.gov/ejscreen/technical-information-about-ejscreen

See also

Examples

head(ej_trend)
#> # A tibble: 6 × 6
#>    year bg           total_pop indicator       value_ptile d2_ptile
#>   <dbl> <chr>            <dbl> <fct>                 <dbl>    <dbl>
#> 1  2018 240010001001       633 lead_paint               44       36
#> 2  2018 240010001001       633 diesel                    8       51
#> 3  2018 240010001001       633 air_cancer                9       48
#> 4  2018 240010001001       633 air_respiratory           8       50
#> 5  2018 240010001001       633 traffic                  38       37
#> 6  2018 240010001001       633 wastewater               82       11