Shapefile of nonresidential areas

An sf object of data from OpenStreetMap of large nonresidential areas in Maryland. There are several census tracts with very few households that coincide with industrial sites, prisons, and other nonresidential properties, and this is an attempt to identify some of them.

Usage

nonres_sf

Format

An sf data frame with 513 rows and 5 variables:

type

Factor. Site type: one of airport, industrial, military, prison, or protected area. These refer to the keys used to retrieve data from OSM, though some may have matched multiple keys (duplicates were removed).

osm_id

Character. The OpenStreetMap ID for the site; can be uesd to retrieve more metadata.

name

Character. The name of the site.

geometry

POLYGON. The boundaries of the site.

is_low_res

Logical. Gives whether the site intersects with a low-residential tract, defined as tracts with fewer than 500 households based on the 2022 ACS.

Source

OpenStreetMap database via the osmdata package, and American Community Survey 2022 5-year estimates via tidycensus.

Examples

head(nonres_sf)
#>               type    osm_id                                         name
#> 107498819   prison 107498819                        Charles Hickey School
#> 13312534    prison  13312534 Federal Correctional Institution, Cumberland
#> 47010857  military  47010857              Naval Support Activity Bethesda
#> 54276982  military  54276982             Warfield Air National Guard Base
#> 114353119 military 114353119                                Camp Fretterd
#> 154081911 military 154081911                        Churchville Test Area
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               geometry
#> 107498819                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             -76.52874, -76.52872, -76.52798, -76.52674, -76.52492, -76.52275, -76.52245, -76.51837, -76.52023, -76.51822, -76.51655, -76.51459, -76.51254, -76.51080, -76.51328, -76.51779, -76.51823, -76.51963, -76.52066, -76.52072, -76.52231, -76.52390, -76.52634, -76.52736, -76.52789, -76.52874, 39.42530, 39.42535, 39.42645, 39.42763, 39.42878, 39.42950, 39.42968, 39.42904, 39.42594, 39.42524, 39.42535, 39.42326, 39.42266, 39.41965, 39.41822, 39.42132, 39.42106, 39.42021, 39.41960, 39.41967, 39.42123, 39.42279, 39.42416, 39.42402, 39.42473, 39.42530
#> 13312534                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               -78.76103, -78.76533, -78.76611, -78.76613, -78.76590, -78.76552, -78.76466, -78.76222, -78.75573, -78.75511, -78.75411, -78.75334, -78.75095, -78.75028, -78.75013, -78.74999, -78.74992, -78.74988, -78.74988, -78.74995, -78.75004, -78.75008, -78.75016, -78.75024, -78.75054, -78.75202, -78.75227, -78.75243, -78.75263, -78.75279, -78.75293, -78.75629, -78.75643, -78.75645, -78.75642, -78.75636, -78.75612, -78.75858, -78.75914, -78.76019, -78.76072, -78.75904, -78.75858, -78.75889, -78.76103, 39.59134, 39.59480, 39.59575, 39.59669, 39.59757, 39.59809, 39.59852, 39.59876, 39.59889, 39.59896, 39.59918, 39.59936, 39.59922, 39.59928, 39.59923, 39.59919, 39.59917, 39.59913, 39.59908, 39.59889, 39.59876, 39.59870, 39.59862, 39.59858, 39.59845, 39.59782, 39.59770, 39.59760, 39.59746, 39.59731, 39.59714, 39.59308, 39.59285, 39.59269, 39.59248, 39.59234, 39.59208, 39.59328, 39.59263, 39.59241, 39.59175, 39.59034, 39.59001, 39.58962, 39.59134
#> 47010857                                                                                                                                                                                                                                                                                                                             -77.09705, -77.09710, -77.09712, -77.09710, -77.09706, -77.09711, -77.09716, -77.09713, -77.09714, -77.09634, -77.09321, -77.09238, -77.09119, -77.09022, -77.08997, -77.08873, -77.08751, -77.08701, -77.08636, -77.08513, -77.08392, -77.08354, -77.08350, -77.08339, -77.08334, -77.08330, -77.08387, -77.08406, -77.08447, -77.08518, -77.08520, -77.08538, -77.08569, -77.08650, -77.08666, -77.08764, -77.08778, -77.08806, -77.08816, -77.08839, -77.08857, -77.08882, -77.08910, -77.08939, -77.09029, -77.09069, -77.09097, -77.09132, -77.09166, -77.09196, -77.09232, -77.09255, -77.09348, -77.09390, -77.09627, -77.09635, -77.09642, -77.09647, -77.09648, -77.09663, -77.09614, -77.09613, -77.09616, -77.09618, -77.09629, -77.09658, -77.09659, -77.09665, -77.09665, -77.09674, -77.09705, 39.00344, 39.00344, 39.00348, 39.00352, 39.00353, 39.00413, 39.00443, 39.00447, 39.00458, 39.00469, 39.00515, 39.00514, 39.00514, 39.00521, 39.00549, 39.00647, 39.00742, 39.00721, 39.00703, 39.00629, 39.00572, 39.00525, 39.00454, 39.00223, 39.00104, 39.00029, 39.00011, 39.00006, 38.99997, 38.99979, 38.99979, 38.99975, 38.99960, 38.99936, 38.99933, 38.99924, 38.99917, 38.99904, 38.99900, 38.99891, 38.99882, 38.99862, 38.99838, 38.99810, 38.99764, 38.99747, 38.99736, 38.99726, 38.99718, 38.99713, 38.99709, 38.99708, 38.99708, 38.99708, 38.99708, 38.99709, 38.99712, 38.99717, 38.99723, 38.99884, 38.99887, 38.99894, 38.99932, 38.99939, 38.99941, 38.99937, 38.99939, 38.99960, 38.99962, 38.99968, 39.00344
#> 54276982                                                                                                                                                                                                                                                                                                                             -76.42237, -76.42198, -76.42163, -76.42121, -76.42085, -76.42017, -76.41954, -76.41892, -76.41868, -76.41843, -76.41763, -76.41631, -76.41534, -76.41543, -76.41544, -76.41462, -76.41463, -76.41355, -76.41248, -76.41117, -76.41031, -76.41007, -76.40983, -76.40943, -76.40963, -76.40981, -76.40982, -76.40967, -76.40937, -76.40925, -76.40926, -76.40914, -76.40888, -76.40823, -76.40779, -76.40767, -76.40756, -76.40744, -76.40713, -76.40644, -76.40583, -76.40554, -76.40511, -76.40467, -76.40390, -76.40357, -76.40321, -76.40316, -76.40321, -76.40351, -76.40371, -76.40429, -76.40557, -76.40604, -76.40917, -76.40925, -76.41148, -76.41431, -76.41631, -76.41858, -76.41902, -76.41942, -76.42213, -76.42367, -76.42389, -76.42385, -76.42382, -76.42383, -76.42410, -76.42376, -76.42237, 39.33698, 39.33706, 39.33710, 39.33714, 39.33714, 39.33709, 39.33699, 39.33678, 39.33668, 39.33657, 39.33626, 39.33570, 39.33528, 39.33421, 39.33412, 39.33404, 39.33316, 39.33302, 39.33286, 39.33290, 39.33277, 39.33214, 39.33210, 39.33108, 39.33049, 39.33037, 39.33011, 39.32981, 39.32952, 39.32904, 39.32857, 39.32822, 39.32814, 39.32706, 39.32682, 39.32670, 39.32615, 39.32586, 39.32555, 39.32524, 39.32437, 39.32399, 39.32380, 39.32374, 39.32350, 39.32343, 39.32316, 39.32297, 39.32255, 39.32232, 39.32161, 39.32117, 39.32214, 39.32220, 39.32479, 39.32500, 39.32668, 39.32816, 39.32973, 39.33149, 39.33184, 39.33215, 39.33428, 39.33549, 39.33560, 39.33575, 39.33586, 39.33598, 39.33647, 39.33657, 39.33698
#> 114353119 -76.84983, -76.84835, -76.84731, -76.84717, -76.84706, -76.84692, -76.84679, -76.84638, -76.84574, -76.84430, -76.84388, -76.84347, -76.84316, -76.84269, -76.84186, -76.84069, -76.84025, -76.83999, -76.83966, -76.83930, -76.83910, -76.83788, -76.83777, -76.83769, -76.83759, -76.83736, -76.83713, -76.83684, -76.83652, -76.83755, -76.83722, -76.83573, -76.83523, -76.83514, -76.83506, -76.83498, -76.83476, -76.83461, -76.83445, -76.83429, -76.83401, -76.83394, -76.83400, -76.83397, -76.83390, -76.83386, -76.83381, -76.83374, -76.83364, -76.83390, -76.83404, -76.83414, -76.83422, -76.83431, -76.83486, -76.83495, -76.83506, -76.83518, -76.83529, -76.83540, -76.83638, -76.83704, -76.83729, -76.83520, -76.83465, -76.83518, -76.83958, -76.84182, -76.84448, -76.84628, -76.84914, -76.85205, -76.85347, -76.85678, -76.85694, -76.85706, -76.85721, -76.85738, -76.85765, -76.85793, -76.86043, -76.86069, -76.85928, -76.86089, -76.85957, -76.84983, 39.50122, 39.50398, 39.50369, 39.50366, 39.50366, 39.50368, 39.50371, 39.50384, 39.50401, 39.50431, 39.50438, 39.50444, 39.50446, 39.50449, 39.50453, 39.50457, 39.50460, 39.50463, 39.50472, 39.50336, 39.50263, 39.50272, 39.50271, 39.50267, 39.50261, 39.50251, 39.50244, 39.50233, 39.50212, 39.50057, 39.50044, 39.50030, 39.50023, 39.50023, 39.50025, 39.50032, 39.50060, 39.50084, 39.50112, 39.50145, 39.50157, 39.50154, 39.50146, 39.50145, 39.50141, 39.50140, 39.50148, 39.50144, 39.50102, 39.50097, 39.50092, 39.50087, 39.50081, 39.50072, 39.50006, 39.50001, 39.49998, 39.49996, 39.49996, 39.49997, 39.50006, 39.49930, 39.49881, 39.49727, 39.49728, 39.49628, 39.49333, 39.49282, 39.49240, 39.49299, 39.49375, 39.49392, 39.49405, 39.49119, 39.49136, 39.49147, 39.49159, 39.49171, 39.49186, 39.49201, 39.49325, 39.49337, 39.49515, 39.49682, 39.49829, 39.50122
#> 154081911                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  -76.25945, -76.25981, -76.25981, -76.25890, -76.25814, -76.25742, -76.25689, -76.25534, -76.25365, -76.25352, -76.25097, -76.25068, -76.25061, -76.25073, -76.25068, -76.25065, -76.25050, -76.25017, -76.25032, -76.24941, -76.24899, -76.24899, -76.24857, -76.24853, -76.24791, -76.24757, -76.24657, -76.24629, -76.24603, -76.24628, -76.24644, -76.24689, -76.24632, -76.24613, -76.24599, -76.24467, -76.24481, -76.24529, -76.24598, -76.24631, -76.24588, -76.24602, -76.24624, -76.24639, -76.24672, -76.24704, -76.24802, -76.24833, -76.24840, -76.24917, -76.24968, -76.25014, -76.25124, -76.25194, -76.25425, -76.25909, -76.25945, 39.59768, 39.60018, 39.60121, 39.60108, 39.60085, 39.60051, 39.60014, 39.59941, 39.59847, 39.59842, 39.59867, 39.59924, 39.59945, 39.59979, 39.60021, 39.60032, 39.60076, 39.60162, 39.60168, 39.60344, 39.60330, 39.60306, 39.60306, 39.60257, 39.60264, 39.60327, 39.60274, 39.60192, 39.60117, 39.60097, 39.60016, 39.59906, 39.59880, 39.59870, 39.59857, 39.59735, 39.59720, 39.59647, 39.59656, 39.59628, 39.59604, 39.59572, 39.59522, 39.59490, 39.59470, 39.59460, 39.59447, 39.59436, 39.59505, 39.59504, 39.59488, 39.59474, 39.59411, 39.59351, 39.59274, 39.59268, 39.59768
#>           is_low_res
#> 107498819      FALSE
#> 13312534       FALSE
#> 47010857       FALSE
#> 54276982       FALSE
#> 114353119      FALSE
#> 154081911      FALSE