Police stops in Connecticut

A dataset containing responses to questions about unfair treatment by police from the 2021 DataHaven Community Wellbeing Survey. The data is broken down by location, category, and group, with values for adults in Connecticut, New Haven, and the Greater New Haven area.

Usage

police_stops

Format

A data frame with 13 rows and 5 variables:

name

Character. The name of the location where the data was collected.

category

Factor. Category for which data are aggregated (total, gender, or race/ethnicity).

group

Factor. The group within the given category for which data are aggregated (Total, Male, Female, White, Black, Latino).

ever_unfairly_stopped

Numeric. The share of respondents who reported having ever been unfairly stopped, harassed, or abused by police.

multiple_times_3yr

Numeric. The share of respondents who reported this treatment having happened multiple times in the past 3 years, out of the share who reported it having happened ever.

Source

DataHaven Community Wellbeing Survey 2021, analyzed with Camille's packages cwi and dcws (not currently public). https://ctdatahaven.org/reports/datahaven-community-wellbeing-survey

Examples

 head(police_stops)
#> # A tibble: 6 × 5
#>   name        category       group  ever_unfairly_stopped multiple_times_3yr
#>   <chr>       <fct>          <fct>                  <dbl>              <dbl>
#> 1 Connecticut Total          Total                   0.15              0.293
#> 2 Connecticut Gender         Male                    0.2               0.27 
#> 3 Connecticut Gender         Female                  0.11              0.354
#> 4 Connecticut Race/Ethnicity White                   0.12              0.162
#> 5 Connecticut Race/Ethnicity Black                   0.25              0.4  
#> 6 Connecticut Race/Ethnicity Latino                  0.2               0.5