Census tracts for the Baltimore area

A dataset containing 2020 census tract boundaries for Baltimore and surrounding counties. The data comes from the tigris package with geometries simplified.

Usage

tracts_sf

Format

An sf data frame with 606 rows and 3 variables:

county

Character. The name of the county where the census tract is located.

geoid

Character. The FIPS code for the census tract.

geometry

POLYGON. The geometric representation of the census tract boundary.

Source

U.S. Census Bureau, TIGER boundary files via the tigris package

Examples

 head(tracts_sf)
#> Simple feature collection with 6 features and 2 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -76.79125 ymin: 39.13982 xmax: -76.58706 ymax: 39.37142
#> Geodetic CRS:  NAD83
#>                county       geoid                       geometry
#> 1      Baltimore city 24510160300 POLYGON ((-76.64612 39.3031...
#> 2      Baltimore city 24510020100 POLYGON ((-76.59047 39.2916...
#> 3    Baltimore County 24005402602 POLYGON ((-76.76732 39.3673...
#> 4      Baltimore city 24510190300 POLYGON ((-76.64696 39.2881...
#> 5      Baltimore city 24510271801 POLYGON ((-76.68336 39.3517...
#> 6 Anne Arundel County 24003730403 POLYGON ((-76.62029 39.1510...