Vacancy rates and correlated housing measures

A dataset containing vacancy rates and other values related to housing for tracts in Baltimore; Stamford, CT; and New Haven, CT. The data comes from the 2022 American Community Survey (ACS). This was updated to include all tracts in Maryland, though Maryland tracts not in Baltimore city will have NA in the city column.

Usage

vacant

Format

A data frame with 1529 rows and 12 variables:

state

Character. The state FIPS code.

city

Character. The city name.

county

Character. The 3-digit county FIPS code.

county_name

Character. Full county name.

geoid

Character. The tract FIPS code.

total_units

Numeric. The total number of housing units.

vacants

Numeric. The number of vacant housing units.

med_rent

Numeric. The median rent for renter-occupied housing units.

med_housing_value

Numeric. The median housing value for owner-occupied housing units.

vacancy_rate

Numeric. The vacancy rate.

area_sqmi

Numeric. The area in square miles.

housing_density

Numeric. The housing density (total units per square mile).

Source

Calculated from US Census Bureau. American Community Survey 2022 5-year estimates. Calculated by Camille with the tidycensus and cwi packages.

Examples

 head(vacant)
#> # A tibble: 6 × 12
#>   state city  county county_name     geoid       total_units vacants med_rent
#>   <chr> <chr> <chr>  <chr>           <chr>             <dbl>   <dbl>    <dbl>
#> 1 24    NA    001    Allegany County 24001000600        1574     200      606
#> 2 24    NA    001    Allegany County 24001001100         857     167      590
#> 3 24    NA    001    Allegany County 24001000800        1033     247      939
#> 4 24    NA    001    Allegany County 24001000100        2245     668      904
#> 5 24    NA    001    Allegany County 24001000200        1617     227      858
#> 6 24    NA    001    Allegany County 24001000700        1769     307      722
#> # ℹ 4 more variables: med_housing_value <dbl>, vacancy_rate <dbl>,
#> #   area_sqmi <dbl>, housing_density <dbl>