Median wages by demographic

A dataset containing median individual earnings by various dimensions (education, occupation, etc.) for the US and Maryland, for adults ages 25 and up with positive earnings. The data is calculated from the 2021 American Community Survey (ACS) Public Use Microdata Sample (PUMS) data via the Integrated Public Use Microdata Series (IPUMS).

Usage

wages

Format

A data frame with 309 rows and 15 variables:

dimension

Factor. The dimension across which values are calculated.

name

Factor. The name of the region (US or Maryland).

sex

Factor. The sex of the individuals.

race_eth

Factor. The race/ethnicity of the individuals.

edu

Factor. The education level of the individuals.

occ_group

Factor. The occupation group of the individuals.

univ

Factor. Universe of workers, whether the denominator is all workers or full-time only.

is_fulltime

Logical. Whether the individual is a full-time worker. NA if distinction isn't included.

count

Numeric. The number of individuals in the group.

earn_q20

Numeric. The 20th percentile of earnings.

earn_q25

Numeric. The 25th percentile of earnings.

earn_q50

Numeric. The 50th percentile (median) earnings.

earn_q75

Numeric. The 75th percentile of earnings.

earn_q80

Numeric. The 80th percentile of earnings.

sample_n

Numeric. The sample size.

Source

U.S. Census Bureau, American Community Survey, Integrated Public Use Microdata Series https://usa.ipums.org/usa/. Analyzed using the srvyr package.

Examples

 head(wages)
#> # A tibble: 6 × 15
#>   dimension name     sex   race_eth edu   occ_group univ      is_fulltime  count
#>   <fct>     <fct>    <fct> <fct>    <fct> <fct>     <fct>     <lgl>        <dbl>
#> 1 total     US       Total Total    Total Total     All work… NA          1.49e8
#> 2 total     Maryland Total Total    Total Total     All work… NA          2.98e6
#> 3 sex       US       Men   Total    Total Total     Full-tim… TRUE        6.02e7
#> 4 sex       Maryland Men   Total    Total Total     Full-tim… TRUE        1.21e6
#> 5 sex       US       Men   Total    Total Total     All work… NA          7.89e7
#> 6 sex       Maryland Men   Total    Total Total     All work… NA          1.53e6
#> # ℹ 6 more variables: earn_q20 <dbl>, earn_q25 <dbl>, earn_q50 <dbl>,
#> #   earn_q75 <dbl>, earn_q80 <dbl>, sample_n <dbl>