Median wages by PUMA

A dataset containing median individual earnings by sex for all Public Use Microdata Areas (PUMAs) in the US, for adults ages 25 and up working full time with positive earnings. The data is calculated from the 2021 American Community Survey (ACS) Public Use Microdata Sample (PUMS) data via the Integrated Public Use Microdata Series (IPUMS). PUMAs are areas of at least 100,000 people.

Usage

wages_by_puma

Format

A data frame with 4702 rows and 5 variables:

statefip

Character. The state FIPS code.

geoid

Character. The PUMA FIPS code.

sex

Character. The sex of the individuals.

count

Numeric. The number of individuals.

earn

Numeric. The median earnings.

Source

U.S. Census Bureau, American Community Survey, Integrated Public Use Microdata Series https://usa.ipums.org/usa/. Analyzed using the srvyr package.

Examples

 head(wages_by_puma)
#> # A tibble: 6 × 5
#>   statefip geoid   sex   count  earn
#>   <chr>    <chr>   <fct> <dbl> <dbl>
#> 1 01       0100100 Women 23434 37092
#> 2 01       0100100 Men   32022 50241
#> 3 01       0100200 Women 28927 50241
#> 4 01       0100200 Men   44487 63586
#> 5 01       0100301 Women 18692 44380
#> 6 01       0100301 Men   27868 72000