Predictions of youth risks in Connecticut

A dataset containing Likert-style responses to questions about different outcomes for youth in their area, from the 2021 DataHaven Community Wellbeing Survey. The data is broken down by category, group, and response with values for adults in Connecticut.

Usage

youth_risks

Format

A data frame with 450 rows and 6 variables:

code

Character. The code for the question.

question

Character. The question about the predicted outcome for youth.

category

Factor. The category of the respondent (e.g., total, gender).

group

Factor. The group of the respondent (e.g., Connecticut, male, female).

response

Factor. The response to the question (almost certain, very likely, a toss up, not very likely, or not at all likely).

value

Numeric. The share of respondents giving the response.

Source

DataHaven Community Wellbeing Survey 2021, analyzed with Camille's packages cwi and dcws (not currently public). https://ctdatahaven.org/reports/datahaven-community-wellbeing-survey

Examples

 head(youth_risks)
#> # A tibble: 6 × 6
#>   code   question                  category group       response           value
#>   <chr>  <chr>                     <fct>    <fct>       <fct>              <dbl>
#> 1 LIFECA Graduate from high school Total    Connecticut Almost certain    0.392 
#> 2 LIFECA Graduate from high school Total    Connecticut Very likely       0.546 
#> 3 LIFECA Graduate from high school Total    Connecticut A toss up         0.0412
#> 4 LIFECA Graduate from high school Total    Connecticut Not very likely   0     
#> 5 LIFECA Graduate from high school Total    Connecticut Not at all likely 0.0103
#> 6 LIFECA Graduate from high school Gender   Male        Almost certain    0.394