Demographics and socio-economic indicators from the 2022 ACS

A dataset containing indicators from the US Census Bureau's 2022 American Community Survey 5-year estimates. These are given for several geographic levels, including the US, every metropolitan statistical area (MSA) in the country, the state of Maryland, and every county and census tract in Maryland.

Usage

acs

Format

A data frame with 1870 rows and 32 variables:

level

Factor. Geographic level (us, msa, state, county, or tract).

county

Character. Name of the county for tracts, NA otherwise.

name

Character. The name of the geography, including FIPS codes for tracts.

total_pop

Numeric. Total population.

white

Numeric. Share of population that is White.

black

Numeric. Share of population that is Black.

latino

Numeric. Share of population that is Latino.

asian

Numeric. Share of population that is Asian.

other_race

Numeric. Share of population that is of other race.

diversity_idx

Numeric. Diversity index based on the preceding race/ethnicity columns. Uses Theil's H entropy.

foreign_born

Numeric. Share of population that is foreign born.

total_hh

Numeric. Total households.

homeownership

Numeric. Homeownership rate.

total_cost_burden

Numeric. Share of households that are cost burdened, based on HUD's standard that housing costs should be no more than 30% of a household's total income.

total_severe_cost_burden

Numeric. Share of households that are severely cost burdened, or paying more than 50% of their income toward housing costs.

owner_cost_burden

Numeric. Share of homeowners that are cost burdened.

owner_severe_cost_burden

Numeric. Share of homeowners that are severely cost burdened.

renter_cost_burden

Numeric. Share of renters that are cost burdened.

renter_severe_cost_burden

Numeric. Share of renters that are severely cost burdened

no_vehicle_hh

Numeric. Share of households without a vehicle.

median_hh_income

Numeric. Median household income in 2022 dollars.

ages25plus

Numeric. Population aged 25 and over.

less_than_high_school

Numeric. Share of population aged 25 and over with less than a high school diploma.

high_school_grad

Numeric. Share of population aged 25 and over with a high school diploma.

some_college_or_aa

Numeric. Share of population aged 25 and over with some college or an associate degree.

bachelors

Numeric. Share of population aged 25 and over with a bachelor's degree.

grad_degree

Numeric. Share of population aged 25 and over with a graduate degree.

pov_status_determined

Numeric. Population for whom poverty status is determined.

poverty

Numeric. Poverty rate, or the share of the population for whom poverty status is determined that lives in a household with income below the federal poverty level.

low_income

Numeric. Low-income rate, or the share of the population for whom poverty status is determined that lives in a household with income below 2 times the federal poverty level.

area_sqmi

Numeric. Land area in square miles.

pop_density

Numeric. Population per square mile.

Source

Calculated from US Census Bureau. American Community Survey 2022 5-year estimates. Calculated by Camille with the tidycensus and cwi packages.

Examples

  head(acs)
#> # A tibble: 6 × 32
#>   level county name  total_pop white black latino asian other_race diversity_idx
#>   <fct> <chr>  <chr>     <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl>         <dbl>
#> 1 us    NA     Unit… 331097593  0.59  0.12   0.19  0.06       0.05         0.744
#> 2 msa   NA     Abil…    176656  0.63  0.07   0.25  0.02       0.03         0.626
#> 3 msa   NA     Akro…    700578  0.78  0.12   0.02  0.03       0.05         0.486
#> 4 msa   NA     Alba…    147716  0.41  0.52   0.03  0.01       0.03         0.598
#> 5 msa   NA     Alba…    128598  0.83  0      0.1   0.01       0.06         0.433
#> 6 msa   NA     Alba…    901729  0.77  0.07   0.06  0.05       0.05         0.532
#> # ℹ 22 more variables: foreign_born <dbl>, total_hh <dbl>, homeownership <dbl>,
#> #   total_cost_burden <dbl>, total_severe_cost_burden <dbl>,
#> #   owner_cost_burden <dbl>, owner_severe_cost_burden <dbl>,
#> #   renter_cost_burden <dbl>, renter_severe_cost_burden <dbl>,
#> #   no_vehicle_hh <dbl>, median_hh_income <dbl>, ages25plus <dbl>,
#> #   less_than_high_school <dbl>, high_school_grad <dbl>,
#> #   some_college_or_aa <dbl>, bachelors <dbl>, grad_degree <dbl>, …