Shapefile of nonresidential areas

An sf object of data from OpenStreetMap of large nonresidential areas in Maryland. There are several census tracts with very few households that coincide with industrial sites, prisons, and other nonresidential properties, and this is an attempt to identify some of them.

Usage

nonres_sf

Format

An sf data frame with 513 rows and 5 variables:

type

Factor. Site type: one of airport, industrial, military, prison, or protected area. These refer to the keys used to retrieve data from OSM, though some may have matched multiple keys (duplicates were removed).

osm_id

Character. The OpenStreetMap ID for the site; can be uesd to retrieve more metadata.

name

Character. The name of the site.

geometry

POLYGON. The boundaries of the site.

is_low_res

Logical. Gives whether the site intersects with a low-residential tract, defined as tracts with fewer than 500 households based on the 2022 ACS.

Source

OpenStreetMap database via the osmdata package, and American Community Survey 2022 5-year estimates via tidycensus.

Examples

head(nonres_sf)
#> Simple feature collection with 6 features and 4 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -78.76613 ymin: 38.99708 xmax: -76.24467 ymax: 39.60344
#> Geodetic CRS:  NAD83
#>               type    osm_id                                         name
#> 107498819   prison 107498819                        Charles Hickey School
#> 13312534    prison  13312534 Federal Correctional Institution, Cumberland
#> 47010857  military  47010857              Naval Support Activity Bethesda
#> 54276982  military  54276982             Warfield Air National Guard Base
#> 114353119 military 114353119                                Camp Fretterd
#> 154081911 military 154081911                        Churchville Test Area
#>                                 geometry is_low_res
#> 107498819 POLYGON ((-76.52874 39.4253...      FALSE
#> 13312534  POLYGON ((-78.76103 39.5913...      FALSE
#> 47010857  POLYGON ((-77.09705 39.0034...      FALSE
#> 54276982  POLYGON ((-76.42237 39.3369...      FALSE
#> 114353119 POLYGON ((-76.84983 39.5012...      FALSE
#> 154081911 POLYGON ((-76.25945 39.5976...      FALSE