Average annual consumer spending

A dataset containing mean amounts of money spent on different categories of goods each year, broken down by US household income quintile. The data comes from the US Census Bureau's annual Consumer Expenditure Survey. Dollar amounts are given for the year reported, not adjusted for inflation.

Usage

spending

Format

A data frame with 202 rows and 12 variables:

year

Numeric. The year of the survey.

item

Character. The category of goods.

qtotal

Numeric. The mean amount spent on the item by all households.

q1

Numeric. The mean amount spent on the item by households in the first (lowest) income quintile.

q2

Numeric. The mean amount spent on the item by households in the second income quintile.

q3

Numeric. The mean amount spent on the item by households in the third income quintile.

q4

Numeric. The mean amount spent on the item by households in the fourth income quintile.

q5

Numeric. The mean amount spent on the item by households in the fifth (highest) income quintile.

l2

Character. The second level category of the item. NA if not applicable.

l3

Character. The third level category of the item. NA if not applicable.

l4

Character. The fourth level category of the item. NA if not applicable.

l5

Character. The fifth level category of the item. NA if not applicable.

Source

US Census Bureau's Consumer Expenditure Survey, available from the Bureau of Labor Statistics https://www.bls.gov/cex/data.htm

Examples

 head(spending)
#> # A tibble: 6 × 12
#>    year item       qtotal    q1    q2    q3    q4     q5 l2    l3    l4    l5   
#>   <dbl> <chr>       <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <chr> <chr> <chr> <chr>
#> 1  2019 Average a…  63036 28672 40472 53045 71173 121571 NA    NA    NA    NA   
#> 2  2019 Food         8169  4400  5859  7505  9080  13987 Food  NA    NA    NA   
#> 3  2019 Food at h…   4643  2790  3672  4422  5198   7129 Food  Food… NA    NA   
#> 4  2019 Cereals a…    583   373   466   549   660    867 Food  Food… Cere… NA   
#> 5  2019 Cereals a…    184   105   142   175   206    289 Food  Food… Cere… Cere…
#> 6  2019 Bakery pr…    400   268   324   374   454    578 Food  Food… Cere… Bake…